refactor(backend/executor): extract billing module, rename extra_runtime_cost
- Create executor/billing.py: all billing logic extracted from ExecutionProcessor
  as module-level free functions (charge_usage, charge_extra_runtime_cost,
  charge_node_usage, handle_post_execution_billing, handle_low_balance,
  handle_insufficient_funds_notif, handle_agent_run_notif, etc.)
- Rename Block.extra_credit_charges → extra_runtime_cost across _base.py,
  orchestrator.py, billing.py and tests
- Rename charge_extra_iterations → charge_extra_runtime_cost
- Remove _charge_and_check_balance: inline logic into charge_node_usage
- ExecutionProcessor keeps thin async delegation methods for public API
  (charge_node_usage, charge_extra_runtime_cost) called by orchestrator.py
- manager.py drops ~500 lines; billing concerns now live in one place
